Move getPath method to TestGraph
